+++ app/program/page.tsx
@@ -1,11 +1,17 @@
-import Container from "app/_components/LayoutsComp/Container";
-import Nav from "app/_components/NavComp/ProjectNav";
+import { Metadata } from "next";
+import HomeArticles from "app/_components/ProgramComp/HomeArticles";
+import HomeAside from "app/_components/ProgramComp/HomeAside";
 
-export default async function Layout(): Promise<JSX.Element> {
+export const metadata: Metadata = {
+  title: `Program || Codinasion`,
+  description: `Share solutions with the community on Codinasion's Program repository. Find good first issues, make coordinated contributions, and join an inclusive community.`,
+};
+
+export default function Page() {
   return (
-    <Container style="pt-20">
-      <Nav layout={{ name: `program`, slug: 'program' }} />
-      <h1 className="w-full text-black py-10 dark:text-white text-center">Components Import Check</h1>
-    </Container>
+    <section className={"flex lg:container mx-auto p-2 gap-3 relative"}>
+      <HomeArticles />
+      <HomeAside />
+    </section>
   )
 }
